remediation/tenant-notifier/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pyyaml==6.0.2
httpx==0.28.1
